+++ CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- ✨(frontend) add duplicate action to doc tree #1175
+
+### Changed
+
+- ♻️(frontend) redirect to doc after duplicate #1175
+
 ### Fixed
 
 - 🌐(frontend) keep simple tag during export #1154